Addressed CR comments.
